// shadcn-admin-main/src/features/tasks/components/tasks-mutate-drawer.tsx
import { z } from 'zod'
import { useForm } from 'react-hook-form'
import { zodResolver } from '@hookform/resolvers/zod'
import { useTranslation } from 'react-i18next'
import { showSubmittedData } from '@/lib/show-submitted-data'
import { Button } from '@/components/ui/button'
import {
  Form,
  FormControl,
  FormField,
  FormItem,
  FormLabel,
  FormMessage,
} from '@/components/ui/form'
import { Input } from '@/components/ui/input'
import { RadioGroup, RadioGroupItem } from '@/components/ui/radio-group'
import {
  Sheet,
  SheetClose,
  SheetContent,
  SheetDescription,
  SheetFooter,
  SheetHeader,
  SheetTitle,
} from '@/components/ui/sheet'
import { SelectDropdown } from '@/components/select-dropdown'
import { type Task } from '../data/schema'

type TaskMutateDrawerProps = {
  open: boolean
  onOpenChange: (open: boolean) => void
  currentRow?: Task
}

export function TasksMutateDrawer({
  open,
  onOpenChange,
  currentRow,
}: TaskMutateDrawerProps) {
  const isUpdate = !!currentRow
  const { t } = useTranslation(['tasks', 'common'])

  const formSchema = z.object({
    title: z.string().min(1, t('tasks:mutateDrawer.validation.titleRequired')),
    status: z
      .string()
      .min(1, t('tasks:mutateDrawer.validation.statusRequired')),
    label: z.string().min(1, t('tasks:mutateDrawer.validation.labelRequired')),
    priority: z
      .string()
      .min(1, t('tasks:mutateDrawer.validation.priorityRequired')),
  })
  type TaskForm = z.infer<typeof formSchema>

  const form = useForm<TaskForm>({
    resolver: zodResolver(formSchema),
    defaultValues: currentRow ?? {
      title: '',
      status: '',
      label: '',
      priority: '',
    },
  })

  const onSubmit = (data: TaskForm) => {
    // do something with the form data
    onOpenChange(false)
    form.reset()
    showSubmittedData(data)
  }

  return (
    <Sheet
      open={open}
      onOpenChange={(v) => {
        onOpenChange(v)
        form.reset()
      }}
    >
      <SheetContent className='flex flex-col'>
        <SheetHeader className='text-start'>
          <SheetTitle>
            {isUpdate
              ? t('tasks:mutateDrawer.titleUpdate')
              : t('tasks:mutateDrawer.titleCreate')}
          </SheetTitle>
          <SheetDescription>
            {isUpdate
              ? t('tasks:mutateDrawer.descUpdate')
              : t('tasks:mutateDrawer.descCreate')}
          </SheetDescription>
        </SheetHeader>
        <Form {...form}>
          <form
            id='tasks-form'
            onSubmit={form.handleSubmit(onSubmit)}
            className='flex-1 space-y-6 overflow-y-auto px-4'
          >
            <FormField
              control={form.control}
              name='title'
              render={({ field }) => (
                <FormItem>
                  <FormLabel>{t('tasks:mutateDrawer.titleField')}</FormLabel>
                  <FormControl>
                    <Input
                      {...field}
                      placeholder={t('tasks:mutateDrawer.titlePlaceholder')}
                    />
                  </FormControl>
                  <FormMessage />
                </FormItem>
              )}
            />
            <FormField
              control={form.control}
              name='status'
              render={({ field }) => (
                <FormItem>
                  <FormLabel>{t('tasks:mutateDrawer.status')}</FormLabel>
                  <SelectDropdown
                    defaultValue={field.value}
                    onValueChange={field.onChange}
                    placeholder={t('tasks:mutateDrawer.statusPlaceholder')}
                    items={[
                      {
                        label: t('tasks:mutateDrawer.statusOptions.inProgress'),
                        value: 'in progress',
                      },
                      {
                        label: t('tasks:mutateDrawer.statusOptions.backlog'),
                        value: 'backlog',
                      },
                      {
                        label: t('tasks:mutateDrawer.statusOptions.todo'),
                        value: 'todo',
                      },
                      {
                        label: t('tasks:mutateDrawer.statusOptions.canceled'),
                        value: 'canceled',
                      },
                      {
                        label: t('tasks:mutateDrawer.statusOptions.done'),
                        value: 'done',
                      },
                    ]}
                  />
                  <FormMessage />
                </FormItem>
              )}
            />
            <FormField
              control={form.control}
              name='label'
              render={({ field }) => (
                <FormItem className='relative'>
                  <FormLabel>{t('tasks:mutateDrawer.label')}</FormLabel>
                  <FormControl>
                    <RadioGroup
                      onValueChange={field.onChange}
                      defaultValue={field.value}
                      className='flex flex-col space-y-1'
                    >
                      <FormItem className='flex items-center'>
                        <FormControl>
                          <RadioGroupItem value='documentation' />
                        </FormControl>
                        <FormLabel className='font-normal'>
                          {t('tasks:mutateDrawer.labelOptions.documentation')}
                        </FormLabel>
                      </FormItem>
                      <FormItem className='flex items-center'>
                        <FormControl>
                          <RadioGroupItem value='feature' />
                        </FormControl>
                        <FormLabel className='font-normal'>
                          {t('tasks:mutateDrawer.labelOptions.feature')}
                        </FormLabel>
                      </FormItem>
                      <FormItem className='flex items-center'>
                        <FormControl>
                          <RadioGroupItem value='bug' />
                        </FormControl>
                        <FormLabel className='font-normal'>
                          {t('tasks:mutateDrawer.labelOptions.bug')}
                        </FormLabel>
                      </FormItem>
                    </RadioGroup>
                  </FormControl>
                  <FormMessage />
                </FormItem>
              )}
            />
            <FormField
              control={form.control}
              name='priority'
              render={({ field }) => (
                <FormItem className='relative'>
                  <FormLabel>{t('tasks:mutateDrawer.priority')}</FormLabel>
                  <FormControl>
                    <RadioGroup
                      onValueChange={field.onChange}
                      defaultValue={field.value}
                      className='flex flex-col space-y-1'
                    >
                      <FormItem className='flex items-center'>
                        <FormControl>
                          <RadioGroupItem value='high' />
                        </FormControl>
                        <FormLabel className='font-normal'>
                          {t('tasks:mutateDrawer.priorityOptions.high')}
                        </FormLabel>
                      </FormItem>
                      <FormItem className='flex items-center'>
                        <FormControl>
                          <RadioGroupItem value='medium' />
                        </FormControl>
                        <FormLabel className='font-normal'>
                          {t('tasks:mutateDrawer.priorityOptions.medium')}
                        </FormLabel>
                      </FormItem>
                      <FormItem className='flex items-center'>
                        <FormControl>
                          <RadioGroupItem value='low' />
                        </FormControl>
                        <FormLabel className='font-normal'>
                          {t('tasks:mutateDrawer.priorityOptions.low')}
                        </FormLabel>
                      </FormItem>
                    </RadioGroup>
                  </FormControl>
                  <FormMessage />
                </FormItem>
              )}
            />
          </form>
        </Form>
        <SheetFooter className='gap-2'>
          <SheetClose asChild>
            <Button variant='outline'>{t('common:close')}</Button>
          </SheetClose>
          <Button form='tasks-form' type='submit'>
            {t('common:saveChanges')}
          </Button>
        </SheetFooter>
      </SheetContent>
    </Sheet>
  )
}
